Backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cssselect2==0.7.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
fonttools==4.55.0
idna==3.10
mysqlclient==2.2.6
numpy==2.1.3
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
